Improved common library functionality.
Moved shared functionality from the continuous and discrete folders to the commons package. This includes the classes CompPair, CompList, DerivativeOrder, and Observable.

Corrected Observable returning ValueError instead of TypeError when attempting to make comparison with objects of the wrong class.

[Irrelevant] PyCharm added metadata to the NS_test notebook. Shouldn't affect functionality.
